Create e4-3.c and update 3-3.c
Added a comment on 3-3.c because I used the function isnumber() but in
retrospect I'm not sure where I got that from. I may have just guessed
it was correct. In general the correct function appears to be isdigit(),
while isnumber() seems to be an Apple-specific function that just so
happened to also work. Nonetheless I've left the code as is.