Fixed 4-4.c and added clearstack()
Fixed existing stack alteration functions so that they actually find the
top of the stack, which is NOT necessarily where sp is. Added the
clearstack() function as well. These functions assume that the array val
is initialized to consist of all zero bits, i.e. all '\0'. This
certainly seems to be the case but I need to reconfirm this is indeed
standard defined behavior for C.